11 epic 3 us 8

+++ src/Controller/ProductController.php
@@ -18,7 +18,7 @@ final class ProductController extends AbstractController
     public function index(ProductRepository $productRepository): Response
     {
         return $this->render('product/index.html.twig', [
-            'products' => $productRepository->findAll(),
+            'products' => $productRepository->findByUser($this->getUser()),
         ]);
     }
 
@@ -30,6 +30,7 @@ final class ProductController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $product->setUser($this->getUser());
             $entityManager->persist($product);
             $entityManager->flush();
 
@@ -46,6 +47,9 @@ final class ProductController extends AbstractController
     #[Route('/{id}', name: 'app_product_show', methods: ['GET'])]
     public function show(Product $product): Response
     {
+        if ($product->getUser() !== $this->getUser()){
+            throw $this->createAccessDeniedException();
+        }
         return $this->render('product/show.html.twig', [
             'product' => $product,
         ]);
@@ -54,6 +58,10 @@ final class ProductController extends AbstractController
     #[Route('/{id}/edit', name: 'app_product_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Product $product, EntityManagerInterface $entityManager): Response
     {
+        if ($product->getUser() !== $this->getUser()){
+            throw $this->createAccessDeniedException();
+        }
+        
         $form = $this->createForm(ProductType::class, $product);
         $form->handleRequest($request);
 
@@ -62,7 +70,7 @@ final class ProductController extends AbstractController
 
             return $this->redirectToRoute('app_product', [], Response::HTTP_SEE_OTHER);
         }
-
+        
         return $this->render('product/edit.html.twig', [
             'product' => $product,
             'form' => $form,
@@ -72,6 +80,10 @@ final class ProductController extends AbstractController
     #[Route('/{id}', name: 'app_product_delete', methods: ['POST'])]
     public function delete(Request $request, Product $product, EntityManagerInterface $entityManager): Response
     {
+        if ($product->getUser() !== $this->getUser()){
+            throw $this->createAccessDeniedException();
+        }
+
         if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($product);
             $entityManager->flush();
